refactor: remove codeBlockOption from default page
