+++ packages/blocks/src/page-block/default/default-page-block.ts
@@ -25,7 +25,6 @@ import type { PageBlockModel } from '../index.js';
 import { bindHotkeys, removeHotkeys } from '../utils/bind-hotkey.js';
 import { deleteModelsByRange, tryUpdateFrameSize } from '../utils/index.js';
 import {
-  CodeBlockOptionContainer,
   DraggingArea,
   EmbedEditingContainer,
   EmbedSelectedRectsContainer,
@@ -43,14 +42,12 @@ export interface EmbedEditingState {
   model: BaseBlockModel;
 }
 
-export type CodeBlockOption = EmbedEditingState;
-
 export interface DefaultSelectionSlots {
   draggingAreaUpdated: Slot<DOMRect | null>;
   selectedRectsUpdated: Slot<DOMRect[]>;
   embedRectsUpdated: Slot<DOMRect[]>;
   embedEditingStateUpdated: Slot<EmbedEditingState | null>;
-  codeBlockOptionUpdated: Slot<CodeBlockOption | null>;
+  codeBlockOptionUpdated?: Slot;
   nativeSelectionToggled: Slot<boolean>;
 }
 
@@ -147,9 +144,6 @@ export class DefaultPageBlockComponent
 
   private _resizeObserver: ResizeObserver | null = null;
 
-  @property()
-  codeBlockOption!: CodeBlockOption | null;
-
   @query('.affine-default-viewport')
   viewportElement!: HTMLDivElement;
 
@@ -158,7 +152,6 @@ export class DefaultPageBlockComponent
     selectedRectsUpdated: new Slot<DOMRect[]>(),
     embedRectsUpdated: new Slot<DOMRect[]>(),
     embedEditingStateUpdated: new Slot<EmbedEditingState | null>(),
-    codeBlockOptionUpdated: new Slot<CodeBlockOption | null>(),
     nativeSelectionToggled: new Slot<boolean>(),
   };
 
@@ -425,10 +418,6 @@ export class DefaultPageBlockComponent
       this._embedEditingState = embedEditingState;
       this.requestUpdate();
     });
-    slots.codeBlockOptionUpdated.on(codeBlockOption => {
-      this.codeBlockOption = codeBlockOption;
-      this.requestUpdate();
-    });
     slots.nativeSelectionToggled.on(flag => {
       if (flag) window.addEventListener('keydown', this._handleNativeKeydown);
       else window.removeEventListener('keydown', this._handleNativeKeydown);
@@ -519,9 +508,6 @@ export class DefaultPageBlockComponent
       this.slots,
       viewport
     );
-    const codeBlockOptionContainer = CodeBlockOptionContainer(
-      page.readonly ? null : this.codeBlockOption
-    );
 
     return html`
       <div class="affine-default-viewport">
@@ -538,7 +524,7 @@ export class DefaultPageBlockComponent
           ${childrenContainer}
         </div>
         ${selectedRectsContainer} ${draggingArea} ${selectedEmbedContainer}
-        ${embedEditingContainer} ${codeBlockOptionContainer}
+        ${embedEditingContainer}
       </div>
     `;
   }
