+++ packages/blocks/src/page-block/index.ts
@@ -1,8 +1,5 @@
-export { CodeBlockOptionContainer } from './default/components.js';
 export * from './default/default-page-block.js';
-export { copyCode, getAllowSelectedBlocks } from './default/utils.js';
-export { deleteCodeBlock } from './default/utils.js';
-export { toggleWrap } from './default/utils.js';
+export { getAllowSelectedBlocks } from './default/utils.js';
 export * from './edgeless/edgeless-page-block.js';
 export * from './page-model.js';
 export * from './utils/index.js';
